Erro: quadras nao estavam sendo salvas. Motivo: o arquivo. geo era lido ate o final para saber a quantidade de quadras, mas ao chamar a funcao "leGeo" o ponteiro para o arquivo continuava no fim, encerrando a funcao sem ler nenhuma linha. Correcao: adicionar "rewind(geo)" antes de chamar a funcao "leGeo" para que o ponteiro retorne para o inicio do arquivo

+++ include/geo.h
@@ -21,6 +21,6 @@ int leNumeroQuadras(FILE* geo);
 /// @param ts Estilo das quadras
 /// @param max_x Maior coordenada x das quadras
 /// @param max_y Maior coordenada y das quadras
-void leGeo(FILE* geo, Hashtable hash_quadras, FILE* svgGeo,Quadra* vetQuadras, Estilo ts, int* max_x, int* max_y);
+void leGeo(FILE* geo, Hashtable hash_quadras, FILE* svgGeo,Quadra* vetQuadras, Estilo ts, double* max_x, double* max_y);
 
 #endif
